+++ NEWS.md
@@ -10,6 +10,8 @@ New
 * **FocusWithSelector string (string...)** action that can be used to
   explicitly try to focus a window. The currently available selectors are:
   _pointer_, _workspacelastofocused_, _top_ and _root_.
+* **--standalone** option to _pekwm_wm_, convenience to improve the debugging
+  experience.
 
 Updated
 -------